Various IRC Fixes and cleanup. Added Vhost handler for on connect. Moved ping command to debug. Removed old debugging messages. Force pydle version to 0.9.1+
